+++ scripts/monitor.sh
@@ -3,4 +3,4 @@
 # Start the Celery monitor.
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-CELERY_LOADER=ewoksjob.config.EwoksLoader CELERY_CONFIG_URI=$SCRIPT_DIR/celeryconfig.py celery flower $@
+EWOKS_CONFIG_URI=$SCRIPT_DIR/celeryconfig.py ewoksjob monitor $@