+++ src/ewoksjob/client/__init__.py
@@ -1,1 +1,3 @@
+"""Client utilities to submit and cancel jobs"""
+
 from .celery import *  # noqa F403