+++ src/ewoksjob/tests/utils.py
@@ -18,17 +18,18 @@ def get_result(future, **kwargs):
         return future.result(**kwargs)
 
 
-def wait_not_finished(mod: ModuleType, expected: set, timeout=3):
+def wait_not_finished(mod: ModuleType, expected_task_ids: set, timeout=3):
+    """Wait until all running task ID's are `expected_task_ids`"""
     if mod.__name__.endswith("celery") and not has_redis_server():
         time.sleep(0.1)
         logger.warning("memory and sqlite do not support task monitoring")
         return
     t0 = time.time()
     while True:
-        task_ids = set(mod.get_not_finished())
-        if task_ids == expected:
+        task_ids = set(mod.get_not_finished_task_ids())
+        if task_ids == expected_task_ids:
             return
         dt = time.time() - t0
         if dt > timeout:
-            assert task_ids == expected, task_ids
+            assert task_ids == expected_task_ids, task_ids
         time.sleep(0.2)