+++ src/ewoksjob/worker/slurm.py
@@ -1,55 +1,94 @@
 """SLURM execution pool."""
 
-import logging
-from typing import Callable, Optional
 import weakref
-from celery.concurrency import thread
+import logging
+from functools import wraps
+from typing import Callable, Optional
+
+from celery.concurrency.gevent import TaskPool as _TaskPool
 
 try:
-    from pyslurmutils.concurrent.futures import SlurmExecutor
+    from pyslurmutils.client import SlurmPythonJobRestClient
 except ImportError:
-    SlurmExecutor = None
+    SlurmPythonJobRestClient = None
 
 __all__ = ("TaskPool",)
 
 logger = logging.getLogger(__name__)
 
-_SLURM_EXECUTOR = None
+
+_SLURM_CLIENT = None
 
 
 def get_executor() -> Optional[Callable]:
     try:
-        return _SLURM_EXECUTOR.submit
+        spawn = _SLURM_CLIENT.spawn
     except (AttributeError, ReferenceError):
+        # TaskPool is not instantiated
         return None
+    return _slurm_execute(spawn)
 
 
-class TaskPool(thread.TaskPool):
+class TaskPool(_TaskPool):
     """SLURM Task Pool."""
 
-    executor_options = dict()
+    EXECUTOR_OPTIONS = dict()
 
     def __init__(self, *args, **kwargs):
-        if SlurmExecutor is None:
+        if SlurmPythonJobRestClient is None:
             raise RuntimeError("requires pyslurmutils")
         super().__init__(*args, **kwargs)
-        self._create_slurm_executor()
+        self._create_slurm_client()
 
     def restart(self):
-        self.slurm_executor.shutdown()
-        self.slurm_executor = None
-        self._create_slurm_executor()
+        self._slurm_client.cleanup()
+        self._slurm_client = None
+        self._create_slurm_client()
 
-    def _create_slurm_executor(self):
-        global _SLURM_EXECUTOR
-        self.slurm_executor = SlurmExecutor(
-            max_workers=self.limit, **self.executor_options
+    def _create_slurm_client(self):
+        global _SLURM_CLIENT
+        self._slurm_client = SlurmPythonJobRestClient(
+            max_workers=self.limit, **self.EXECUTOR_OPTIONS
         )
-        _SLURM_EXECUTOR = weakref.proxy(self.slurm_executor)
+        _SLURM_CLIENT = weakref.proxy(self._slurm_client)
 
     def on_stop(self):
-        self.slurm_executor.shutdown()
+        self._slurm_client.cleanup()
+        self._slurm_client = None
         super().on_stop()
 
     def terminate_job(self, pid, signal=None):
-        raise NotImplementedError("SLURM job termination not implemented yet")
+        print("TODO: support job cancelling for the slurm pool")
+
+
+def _slurm_execute(spawn: Callable) -> Callable:
+    @wraps(spawn)
+    def executor(celery_task, *args, **kwargs):
+        from ..apps import ewoks
+
+        ewoks_task = ewoks.get_ewoks_task_from_celery_task(celery_task)
+        if ewoks.celery_task_is_bound(celery_task):
+            args = args[1:]  # remove `self`
+
+        future = spawn(ewoks_task, args=args, kwargs=kwargs)
+        try:
+            return future.result()
+        except BaseException:
+            future.client.log_stdout_stderr(
+                future.job_id, logger=logger, level=logging.ERROR
+            )
+            raise
+        else:
+            future.client.log_stdout_stderr(
+                future.job_id, logger=logger, level=logging.INFO
+            )
+        finally:
+            try:
+                status = future.job_status()
+                logger.info("Slurm job %s, %s", future.job_id, status)
+                if future.job_status() not in ("COMPLETED", "CANCELLED", "FAILED"):
+                    future.cancel_job()
+            finally:
+                future.cleanup_job()
+
+    return executor